Refactor messages and routes
