+++ server/src/rest/routes/index.ts
@@ -1,0 +1,4 @@
+export { root } from './root';
+export { users } from './users';
+export { donations } from './donations';
+export { transactions } from './transactions';
