+++ server/src/rest/routes/transactions.ts
@@ -1,5 +1,5 @@
 import { Router } from 'express';
-import { wrapResponse, requireAuth } from './middleware';
+import { wrapResponse, requireAuth } from '../middleware';
 
 export const transactions = Router();
